Extract StringExtractor+StringCatalog.swift from StringExtractor.swift
